+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.16.0 (2018-XX-XX)
 ....................
 
+* add additional fields to ``Schema`` class to declare validation for ``str`` and numeric values, #311 by @tiangolo
 * refactor schema generation to be compatible with JSON Schema and OpenAPI specs, #308 by @tiangolo
 * add ``schema`` to ``schema`` module to generate top-level schemas from base models, #308 by @tiangolo
 * add ``case_insensitive`` option to ``BaseSettings`` ``Config``, #277 by @jasonkuhrt